requirements.txt
pypdf
python-docx
pandas
pyyaml
scikit-learn
xgboost
streamlit
